requirements.txt
uvicorn
fastapi
jinja2
pymongo
python-dotenv
